Move encoder components
