+++ src/ImageSharp/Formats/Jpeg/GolangPort/JpegEncoderCore.cs
@@ -4,7 +4,7 @@
 using System.IO;
 using System.Runtime.CompilerServices;
 using SixLabors.ImageSharp.Formats.Jpeg.Common;
-using SixLabors.ImageSharp.Formats.Jpeg.GolangPort.Components;
+using SixLabors.ImageSharp.Formats.Jpeg.Components.Encoder;
 using SixLabors.ImageSharp.Formats.Jpeg.GolangPort.Components.Encoder;
 using SixLabors.ImageSharp.MetaData.Profiles.Exif;
 using SixLabors.ImageSharp.MetaData.Profiles.Icc;
